+++ projects/api-client/src/api/testResult.service.ts
@@ -62,7 +62,7 @@ export class TestResultService {
 
 
     /**
-     * Post test result data
+     * Get all test result details for specified build
      * 
      * @param buildId Build ID
      * @param files Test result file
@@ -94,7 +94,6 @@ export class TestResultService {
 
         // to determine the Content-Type header
         const consumes: string[] = [
-            'multipart/form-data'
         ];
 
         const canConsumeForm = this.canConsumeForm(consumes);
@@ -127,7 +126,7 @@ export class TestResultService {
     }
 
     /**
-     * Get all test result details for specified build
+     * Get test result list summary of all tests for specified build
      * 
      * @param buildId Build ID
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
@@ -167,7 +166,7 @@ export class TestResultService {
     }
 
     /**
-     * Get test result list summary of all tests for specified build
+     * Post test result data
      * 
      * @param buildId Build ID
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
@@ -194,6 +193,7 @@ export class TestResultService {
 
         // to determine the Content-Type header
         const consumes: string[] = [
+            'multipart/form-data'
         ];
 
         return this.httpClient.get<ResponseTestResultListSummary>(`${this.basePath}/build/${encodeURIComponent(String(buildId))}/test-result/list-summary`,